+++ index.js
@@ -1,3 +1,5 @@
 import './modules/now.js';
-import './modules/functions.js';
-import './modules/contents.js';
+import './modules/headerLinks.js';
+import DisplayBooks from './modules/contents.js';
+
+DisplayBooks.displaybooks();
